diploma_disk/backend/finddotemployee/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
drf-yasg
eventlet
gunicorn
inflection
packaging
pytz
PyYAML
sqlparse
typing_extensions
tzdata
uritemplate
pillow
psycopg2-binary
PyJWT